feat(combat): added ParticipantDTO and extend CombatResultDTO fields
