+++ app/src/main/java/com/arcathoria/combat/dto/ParticipantDTO.java
@@ -1,0 +1,9 @@
+package com.arcathoria.combat.dto;
+
+import java.util.UUID;
+
+public record ParticipantDTO(
+        UUID id,
+        int currentHp
+) {
+}
